fix: slider components
